pyproject.toml
[tool.poetry]
name = "deploy-matbakhi"
version = "0.1.0"
description = ""
authors = ["yusuf-74 <yachour31@gmail.com>"]
readme = "README.md"
packages = [{include = "deploy_matbakhi"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
langchain = "^0.0.117"
openai = "^0.27.2"
unstructured = "^0.5.4"
chromadb = "^0.3.11"
python-decouple = "^3.8"
django-extensions = "^3.2.1"
twilio = "^7.16.5"
bs4 = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
